Fix imports & clean up Main.scala imports, move config.exceptions
