+++ src/main/scala/io/viash/Main.scala
@@ -17,28 +17,19 @@
 
 package io.viash
 
-import java.io.File
-import java.io.FileNotFoundException
-import java.nio.file.NoSuchFileException
-import java.nio.file.Paths
-import java.nio.file.FileSystemNotFoundException
+import java.io.{File, FileNotFoundException}
+import java.nio.file.{Path, Paths, Files, NoSuchFileException, FileSystemNotFoundException}
+import java.net.URI
 import sys.process.{Process, ProcessLogger}
 
 import config.Config
-import helpers.IO
+import helpers.{IO, Exec, SysEnv}
 import helpers.Scala._
-import cli.{CLIConf, ViashCommand, ViashNs, ViashNsBuild}
-import io.viash.exceptions.{MissingResourceFileException, ExitException, MalformedInputException}
-import io.viash.helpers.status._
-import io.viash.platforms.Platform
-import io.viash.project.ViashProject
-import io.viash.cli.DocumentedSubcommand
-import java.nio.file.Path
-import io.viash.helpers.Exec
-import java.nio.file.Files
-import java.net.URI
-import io.viash.config.AbstractConfigException
-import io.viash.helpers.{ExitException, SysEnv}
+import helpers.status._
+import platforms.Platform
+import project.ViashProject
+import cli.{CLIConf, ViashCommand, DocumentedSubcommand, ViashNs, ViashNsBuild}
+import exceptions._
 
 object Main {
   private val pkg = getClass.getPackage
